Move Loop class * Add file for common code between models
